@@ -1,8 +1,6 @@
-from unittest import mock
-from unittest import skip
+from unittest import mock, skip
 
 from mopidy import backend as backend_api
-
 from mopidy_spotify import backend, library, playlists
 from mopidy_spotify.backend import SpotifyPlaybackProvider
 
@@ -102,9 +100,7 @@ def test_on_start_refreshes_playlists(web_mock, config, caplog):
     assert backend.playlists._loaded
 
 
-def test_on_start_doesnt_refresh_playlists_if_not_allowed(
-    web_mock, config, caplog
-):
+def test_on_start_doesnt_refresh_playlists_if_not_allowed(web_mock, config, caplog):
     config["spotify"]["allow_playlists"] = False
 
     backend = get_backend(config)
