@@ -35,48 +35,55 @@ _PLAYLISTS_DIR_CONTENTS = [
 ]
 
 
-BROWSE_DIR_URIS = set(
+BROWSE_DIR_URIS = {
     u.uri
-    for u in [ROOT_DIR]
-    + _ROOT_DIR_CONTENTS
-    + _TOP_LIST_DIR_CONTENTS
-    + _YOUR_MUSIC_DIR_CONTENTS
-    + _PLAYLISTS_DIR_CONTENTS
-)
+    for u in [
+        ROOT_DIR,
+        *_ROOT_DIR_CONTENTS,
+        *_TOP_LIST_DIR_CONTENTS,
+        *_YOUR_MUSIC_DIR_CONTENTS,
+        *_PLAYLISTS_DIR_CONTENTS,
+    ]
+}
 
 
-def browse(*, config, session, web_client, uri):
+def browse(  # noqa: C901, PLR0911, PLR0912
+    *,
+    config,  # noqa: ARG001
+    session,  # noqa: ARG001
+    web_client,
+    uri,
+):
     if uri == ROOT_DIR.uri:
         return _ROOT_DIR_CONTENTS
-    elif uri == _TOP_LIST_DIR.uri:
+    if uri == _TOP_LIST_DIR.uri:
         return _TOP_LIST_DIR_CONTENTS
-    elif uri == _YOUR_MUSIC_DIR.uri:
+    if uri == _YOUR_MUSIC_DIR.uri:
         return _YOUR_MUSIC_DIR_CONTENTS
-    elif uri == _PLAYLISTS_DIR.uri:
+    if uri == _PLAYLISTS_DIR.uri:
         return _PLAYLISTS_DIR_CONTENTS
 
     if web_client is None or not web_client.logged_in:
         return []
 
     # TODO: Support for category browsing.
-    if uri.startswith("spotify:user:") or uri.startswith("spotify:playlist:"):
+    if uri.startswith(("spotify:user:", "spotify:playlist:")):
         return _browse_playlist(web_client, uri)
-    elif uri.startswith("spotify:album:"):
+    if uri.startswith("spotify:album:"):
         return _browse_album(web_client, uri)
-    elif uri.startswith("spotify:artist:"):
+    if uri.startswith("spotify:artist:"):
         return _browse_artist(web_client, uri)
-    elif uri.startswith("spotify:top:"):
+    if uri.startswith("spotify:top:"):
         parts = uri.replace("spotify:top:", "").split(":")
-        if len(parts) == 1:
-            return _browse_toplist_user(web_client, variant=parts[0])
-        else:
+        if len(parts) != 1:
             logger.info(f"Failed to browse {uri!r}: Toplist URI parsing failed")
             return []
-    elif uri.startswith("spotify:your:"):
+        return _browse_toplist_user(web_client, variant=parts[0])
+    if uri.startswith("spotify:your:"):
         parts = uri.replace("spotify:your:", "").split(":")
         if len(parts) == 1:
             return _browse_your_music(web_client, variant=parts[0])
-    elif uri.startswith("spotify:playlists:"):
+    if uri.startswith("spotify:playlists:"):
         parts = uri.replace("spotify:playlists:", "").split(":")
         if len(parts) == 1:
             return _browse_playlists(web_client, variant=parts[0])
@@ -86,7 +93,12 @@ def browse(*, config, session, web_client, uri):
 
 
 def _browse_playlist(web_client, uri):
-    return playlists.playlist_lookup(web_client, uri, None, as_items=True)
+    return playlists.playlist_lookup(
+        web_client,
+        uri,
+        bitrate=None,
+        as_items=True,
+    )
 
 
 def _browse_album(web_client, uri):
@@ -132,12 +144,11 @@ def _browse_toplist_user(web_client, variant):
                 if page
             ]
         )
-        if variant == "tracks":
-            return list(
-                translator.web_to_track_refs(items, check_playable=False)
-            )
-        else:
-            return list(translator.web_to_artist_refs(items))
+        match variant:
+            case "tracks":
+                return list(translator.web_to_track_refs(items, check_playable=False))
+            case "artists":
+                return list(translator.web_to_artist_refs(items))
     else:
         return []
 
@@ -157,35 +168,35 @@ def _load_your_music(web_client, variant):
         if not page:
             continue
         items = page.get("items", [])
-        for item in items:
-            yield item
+        yield from items
 
 
 def _browse_your_music(web_client, variant):
     items = _load_your_music(web_client, variant)
-    if variant == "tracks":
-        return list(translator.web_to_track_refs(items))
-    elif variant == "albums":
-        return list(translator.web_to_album_refs(items))
-    else:
-        return []
+    match variant:
+        case "tracks":
+            return list(translator.web_to_track_refs(items))
+        case "albums":
+            return list(translator.web_to_album_refs(items))
+        case _:
+            return []
 
 
 def _browse_playlists(web_client, variant):
     if not web_client.logged_in:
         return []
 
-    if variant == "featured":
-        items = flatten(
-            [
-                page.get("playlists", {}).get("items", [])
-                for page in web_client.get_all(
+    if variant != "featured":
+        return []
+
+    items = flatten(
+        [
+            page.get("playlists", {}).get("items", [])
+            for page in web_client.get_all(
                 "browse/featured-playlists",
                 params={"limit": 50},
             )
-                if page
-            ]
-        )
-        return list(translator.to_playlist_refs(items))
-    else:
-        return []
+            if page
+        ]
+    )
+    return list(translator.to_playlist_refs(items))
