@@ -3,8 +3,8 @@ import logging
 import operator
 import urllib.parse
 
-from mopidy import models
 from mopidy_spotify.browse import BROWSE_DIR_URIS
+from mopidy_spotify.translator import web_to_image
 
 _API_MAX_IDS_PER_REQUEST = 50
 
@@ -36,48 +36,50 @@ def get_images(web_client, uris):
 
 def _parse_uri(uri):
     if uri in BROWSE_DIR_URIS:
-        return  # These are internal to the extension.
+        return None  # These are internal to the extension.
     try:
         parsed_uri = urllib.parse.urlparse(uri)
         uri_type, uri_id = None, None
 
-        if parsed_uri.scheme == "spotify":
-            fragments = parsed_uri.path.split(":")
-            if len(fragments) < 2:
-                raise ValueError("Too few fragments")
-            uri_type, uri_id = parsed_uri.path.split(":")[:2]
-        elif parsed_uri.scheme in ("http", "https"):
-            if parsed_uri.netloc in ("open.spotify.com", "play.spotify.com"):
-                uri_type, uri_id = parsed_uri.path.split("/")[1:3]
+        match parsed_uri.scheme:
+            case "spotify":
+                match parsed_uri.path.split(":"):
+                    case uri_type, uri_id, *_:
+                        pass
+                    case _:
+                        raise ValueError("Too few arguments")  # noqa: TRY301
+            case "http" | "https":
+                if parsed_uri.netloc in ("open.spotify.com", "play.spotify.com"):
+                    uri_type, uri_id = parsed_uri.path.split("/")[1:3]
 
         supported_types = ("track", "album", "artist", "playlist")
         if uri_type:
             if uri_type not in supported_types:
-                logger.warning(
-                    f"Unsupported image type '{uri_type}' in {repr(uri)}"
-                )
-                return
-            elif uri_id:
+                logger.warning(f"Unsupported image type '{uri_type}' in {uri!r}")
+                return None
+            if uri_id:
                 return {
                     "uri": uri,
                     "type": uri_type,
                     "id": uri_id,
                     "key": (uri_type, uri_id),
                 }
-        raise ValueError("Unknown error")
+        raise ValueError("Unknown error")  # noqa: TRY301
     except Exception as e:
-        logger.exception(f"Could not parse {repr(uri)} as a Spotify URI ({e})")
+        logger.exception(f"Could not parse {uri!r} as a Spotify URI ({e!s})")  # noqa: TRY401
 
 
 def _process_uri(web_client, uri):
     data = web_client.get(f"{uri['type']}s/{uri['id']}")
-    _cache[uri["key"]] = tuple(
-        _translate_image(i) for i in data.get("images") or []
-    )
+    _cache[uri["key"]] = tuple(web_to_image(i) for i in data.get("images") or [])
     return {uri["uri"]: _cache[uri["key"]]}
 
 
-def _process_uris(web_client, uri_type, uris):
+def _process_uris(  # noqa: C901
+    web_client,
+    uri_type,
+    uris,
+):
     result = {}
     ids = [u["id"] for u in uris]
     ids_to_uris = {u["id"]: u for u in uris}
@@ -113,18 +115,13 @@ def _process_uris(web_client, uri_type, uris):
                 album_key = album["key"]
                 if album_key not in _cache:
                     _cache[album_key] = tuple(
-                        _translate_image(i)
-                        for i in item["album"].get("images") or []
+                        web_to_image(i) for i in item["album"].get("images") or []
                     )
                 _cache[uri["key"]] = _cache[album_key]
             else:
                 _cache[uri["key"]] = tuple(
-                    _translate_image(i) for i in item.get("images") or []
+                    web_to_image(i) for i in item.get("images") or []
                 )
         result[uri["uri"]] = _cache[uri["key"]]
 
     return result
-
-
-def _translate_image(i):
-    return models.Image(uri=i["url"], height=i["height"], width=i["width"])
