@@ -1,27 +1,26 @@
-import collections
 import logging
+from collections.abc import Hashable
 
 from mopidy import models
 
 logger = logging.getLogger(__name__)
 
 
-class memoized:  # noqa N801
+class memoized:  # noqa: N801
     def __init__(self, func):
         self.func = func
         self.cache = {}
 
     def __call__(self, *args, **kwargs):
         # NOTE Only args, not kwargs, are part of the memoization key.
-        if not isinstance(args, collections.abc.Hashable):
+        if not isinstance(args, Hashable):
             return self.func(*args, **kwargs)
         if args in self.cache:
             return self.cache[args]
-        else:
-            value = self.func(*args, **kwargs)
-            if value is not None:
-                self.cache[args] = value
-            return value
+        value = self.func(*args, **kwargs)
+        if value is not None:
+            self.cache[args] = value
+        return value
 
 
 # TODO: memoize web functions?
@@ -29,7 +28,7 @@ class memoized:  # noqa N801
 
 def web_to_artist_ref(web_artist):
     if not valid_web_data(web_artist, "artist"):
-        return
+        return None
 
     uri = web_artist["uri"]
     return models.Ref.artist(uri=uri, name=web_artist.get("name", uri))
@@ -44,7 +43,7 @@ def web_to_artist_refs(web_artists):
 
 def web_to_album_ref(web_album):
     if not valid_web_data(web_album, "album"):
-        return
+        return None
 
     if "name" in web_album:
         artists = web_album.get("artists", [])
@@ -60,9 +59,10 @@ def web_to_album_ref(web_album):
 
 def web_to_album_refs(web_albums):
     for web_album in web_albums:
-        # The extra level here is to also support "saved album objects".
-        web_album = web_album.get("album", web_album)
-        ref = web_to_album_ref(web_album)
+        ref = web_to_album_ref(
+            # The extra level here is to also support "saved album objects".
+            web_album.get("album", web_album),
+        )
         if ref is not None:
             yield ref
 
@@ -77,7 +77,7 @@ def valid_web_data(data, object_type):
 
 def web_to_track_ref(web_track, *, check_playable=True):
     if not valid_web_data(web_track, "track"):
-        return
+        return None
 
     # Web API track relinking guide says to use original URI.
     # libspotfy will handle any relinking when track is loaded for playback.
@@ -85,22 +85,25 @@ def web_to_track_ref(web_track, *, check_playable=True):
 
     if check_playable and not web_track.get("is_playable", False):
         logger.debug(f"{uri!r} is not playable")
-        return
+        return None
 
     return models.Ref.track(uri=uri, name=web_track.get("name", uri))
 
 
 def web_to_track_refs(web_tracks, *, check_playable=True):
     for web_track in web_tracks:
-        # The extra level here is to also support "saved track objects".
-        web_track = web_track.get("track", web_track)
-        ref = web_to_track_ref(web_track, check_playable=check_playable)
+        ref = web_to_track_ref(
+            # The extra level here is to also support "saved track objects".
+            web_track.get("track", web_track),
+            check_playable=check_playable,
+        )
         if ref is not None:
             yield ref
 
 
 def to_playlist(
     web_playlist,
+    *,
     username=None,
     bitrate=None,
     as_ref=False,
@@ -112,7 +115,7 @@ def to_playlist(
 
     web_tracks = web_playlist.get("tracks", {}).get("items") or []
     if as_items and not isinstance(web_tracks, list):
-        return
+        return None
 
     if as_items:
         return list(web_to_track_refs(web_tracks))
@@ -128,7 +131,7 @@ def to_playlist(
 
 def to_playlist_ref(web_playlist, username=None):
     if not valid_web_data(web_playlist, "playlist"):
-        return
+        return None
 
     name = web_playlist.get("name", web_playlist["uri"])
 
@@ -156,19 +159,19 @@ SEARCH_FIELD_MAP = {
 }
 
 
-def sp_search_query(query, exact=False):
+def sp_search_query(query, *, exact=False):
     """Translate a Mopidy search query to a Spotify search query"""
 
     result = []
 
     for field, values in query.items():
-        field = SEARCH_FIELD_MAP.get(field, field)
+        field = SEARCH_FIELD_MAP.get(field, field)  # noqa: PLW2901
         if field is None:
             continue
 
         for value in values:
             if field == "year":
-                value = _transform_year(value)
+                value = _transform_year(value)  # noqa: PLW2901
                 if value is not None:
                     result.append(f"{field}:{value}")
             elif field == "any":
@@ -176,13 +179,10 @@ def sp_search_query(query, exact=False):
                     result.append(f'"{value}"')
                 else:
                     result.append(value)
+            elif exact:
+                result.append(f'{field}:"{value}"')
             else:
-                if exact:
-                    result.append(f'{field}:"{value}"')
-                else:
-                    result.append(
-                        " ".join(f"{field}:{word}" for word in value.split())
-                    )
+                result.append(" ".join(f"{field}:{word}" for word in value.split()))
 
     return " ".join(result)
 
@@ -191,15 +191,13 @@ def _transform_year(date):
     try:
         return int(date.split("-")[0])
     except ValueError:
-        logger.debug(
-            f'Excluded year from search query: Cannot parse date "{date}"'
-        )
+        logger.debug(f'Excluded year from search query: Cannot parse date "{date}"')
 
 
 def web_to_artist(web_artist):
     ref = web_to_artist_ref(web_artist)
     if ref is None:
-        return
+        return None
 
     return models.Artist(uri=ref.uri, name=ref.name)
 
@@ -216,20 +214,16 @@ def web_to_album_tracks(web_album, bitrate=None):
     if not isinstance(web_tracks, list):
         return []
 
-    tracks = [
-        web_to_track(web_track, bitrate, album) for web_track in web_tracks
-    ]
+    tracks = [web_to_track(web_track, bitrate, album) for web_track in web_tracks]
     return [t for t in tracks if t]
 
 
 def web_to_album(web_album):
     ref = web_to_album_ref(web_album)
     if ref is None:
-        return
+        return None
 
-    artists = [
-        web_to_artist(web_artist) for web_artist in web_album.get("artists", [])
-    ]
+    artists = [web_to_artist(web_artist) for web_artist in web_album.get("artists", [])]
     artists = [a for a in artists if a]
 
     name = web_album.get("name", "Unknown album")
@@ -239,16 +233,15 @@ def web_to_album(web_album):
 def int_or_none(inp):
     if inp is not None:
         return int(float(inp))
+    return None
 
 
 def web_to_track(web_track, bitrate=None, album=None):
     ref = web_to_track_ref(web_track)
     if ref is None:
-        return
+        return None
 
-    artists = [
-        web_to_artist(web_artist) for web_artist in web_track.get("artists", [])
-    ]
+    artists = [web_to_artist(web_artist) for web_artist in web_track.get("artists", [])]
     artists = [a for a in artists if a]
 
     if album is None:
@@ -263,4 +256,12 @@ def web_to_track(web_track, bitrate=None, album=None):
         disc_no=int_or_none(web_track.get("disc_number")),
         track_no=int_or_none(web_track.get("track_number")),
         bitrate=bitrate,
+    )
+
+
+def web_to_image(web_image):
+    return models.Image(
+        uri=web_image["url"],
+        height=int_or_none(web_image.get("height")),
+        width=int_or_none(web_image.get("width")),
     )
